chore: change the seed and init logic
With this naive solution, the schema is mounted first, then the seeds. Later on this can easily get adapted to include migrations. The whole migrations folder will be mounted and the schema.sql file and seed.sql file will each be mounter, the former first with a 00000 prefix and the later with a 999999 prefix will run last